requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
tqdm
pyyaml
flask
